Fix tracking_column regression with Postgresql Numeric types
Prior to logsatash-input-jdbc 4.3.14, it was possible to use
the Postgresql Numeric type as a tracking column across Logstash
runs. Since 4.3.14, the deserialization of the sql_last_run
parameter has been broken for BigDecimal types, which is how
Numeric types are represented as Ruby BigDecimal objects by the
Sequel library we use in this plugin.

This commit restores this functionality.

Fixes: https://github.com/logstash-plugins/logstash-input-jdbc/issues/371