Fix prepared statement reconnect issue
This commit fixes an issue where, if the connection between Logstash
and the database fails due to an attempt to use a closed prepared statement.
This results in subsequent attempts to use a prepared statement
failing when using scheduling, which can only be rectified by restarting Logstash.

This commit resets the `statement_prepared` flag, which means that the
prepared statement will be recreated upon the next scheduled run.

This commit also adds a backtrace to log entries when this issue occurs to provide
more visibility into the cause of the issue.