Add extra safety to local data retrieval
Under certain cirumstances, retrieving local data from the embedded derby database
can result in an exception being thrown from the Sequel library that is not wrapped in
a `SequelError`. Prior to this commit, that would result in the exception "escaping" the
plugin, causing the plugin (and the pipeline) to crash.

This commit broadens the scope of the rescue to catch any errors that might result from
that call, enabling the appropriate logging and tagging to take place without crashing
the plugin.

Fixes #77