Switch pattern of usage Sequel to leverage the DB connection pool
This commit changes the way JDBC input plugin uses the Sequel DB access library.
Instead of instantiating a new pool for every query, it creates the pool during registration of the plugin,
then reuse that for every DB access. This avoid the connection leaked happening in scheduled queries,
as described in #118.