Sequel datetime class consistency (#151)
* timezones: absent instruction otherwise, interpret times as local

these tests SUCCEED when run on their own, but FAIL when run after any
instance of a plugin has been initialized with `jdbc_default_timezone`,
since doing so globally configures Sequel to use DateTime instead of
Time:

To validate, modify the `.ci/run.sh` execution line to be:
> ~~~
> bundle exec rspec spec --format documentation --example "without jdbc_default_timezone"
> ~~~

* timezones: consistently pre-load Sequel, force datetime_class=Time

By default, when a plugin is configured _without_ `jdbc_default_timezone`
timestamps are assumed to be in the same timezone as the Logstash host
machine, because they are parsed with Ruby's `Time#parse` which uses
local context.

However, when any one plugin declares `jdbc_default_timezone`, we load
Sequel's `named_timezones` extension, which has a side-effect of globally
changing `Sequel.datetime_class` to ruby's `DateTime`. The plugins that
are configured with `jdbc_default_timezone` have enough information to
apply the separately-provided offset, but the plugins that do _not_ have
a `jdbc_default_timezone` directive become broken and effectively
fail to parse the timestamps as they did when run on their own.

Sequel's `named_timezones` extension supports being used with `Time`
objects, and is noted to override `Sequel#datetime_class` on load only
for historic reasons.

We force Sequel to use ruby's `Time` class globally (enforcing its
default and preventing it from being changed). This is done inside
of a separately-required bootstrap, which allows us to ensure it is
loaded exacltly once.

* bump version & add changelog entry