input: fix issue where retried query could prevent pipeline shutdown
The use of the `retry` keyword when a lock is acquired in the `begin...` block
and released in the `ensure` block will cause the lock to be acquired more
times than released whenever the block is retried. Because only one thread
ever executes the queries, this isn't a problem at runtime, but can cause the
plugin to stall during shutdown when a separate thread attempts to acquire
the lock for cleanup.